HTTP 500 response when tests fail
This is not necessarily backwards compatible (I know I've written tests
on apps that expect the smoketest endpoint to always return a 200), but
I believe is the Right Thing to do to stay in line with REST
expectations. As I'm dealing with more tools that can monitor a
health-check/smoketest endpoint (consul, AWS stuff, etc), the common
approach seems to be to view anything other than a 200 as a fail and not
need to inspect the actual payload. django-smoketest is nice in that it
will give much more detailed output on which tests failed, etc. but it
seems that tools not written explicitly for that just go off the plain
HTTP status code, so we need to support that as well.

Anyway, since it's not backwards compatible, this will necessitate a
major version bump to 1.0.0.